+++ esibd/plugins_internal/RSPD3303C/RSPD3303C.py
@@ -1,4 +1,4 @@
-# pylint: disable=[missing-module-docstring] # only single class in module
+# pylint: disable=[missing-module-docstring] # see class docstrings
 from threading import Thread
 import time
 from random import choices