+++ docs/pages/changelog.rst
@@ -24,6 +24,7 @@ Changed
 - Improved the definition of the channel properties active and real in the documentation.
 - Treating file types not case sensitive.
 - New icons for msScan and MS.
+- Simplified TIC plugin implementation by inheriting from OMNICONTROL.
 
 
 Version 0.7.2 2025-03-02